Crate rattler_menuinst

Modules

schema

Enums

MenuInstError

Functions

install_menuitems

install_menuitems_for_record

Install menu items for a given prefix record according to Menu/*.json files Note: this function will update the prefix record with the installed menu items and write it back to the prefix record file if any Menu item is found

is_menu_schema_path

Returns true if the given relative path points to a Menu schema JSON file within a package (i.e. Menu/*.json).

remove_menu_items

Remove menu items from a given schema file

remove_menuitems_for_record

Remove all menu items from a given prefix record. This function will remove the menu items from the system, and update the prefix record with removing the tracker entries