rattler_menuinst

Module schema

Structs

BasePlatformSpecific

A platform-specific menu item configuration.

CFBundleDocumentTypesModel

Describes a document type associated with the app.

CFBundleURLTypesModel

Describes a URL scheme associated with the app.

Linux

Linux-specific instructions. Check the Desktop entry specification <https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html#recognized-keys> for more details.

MacOS

macOS specific fields in the menuinst. For more information on the keys, read the following URLs

MacOSVersion

macOS version number.

MenuInstSchema

Metadata required to create menu items across operating systems with menuinst

MenuItem

MenuItemCommand

Instructions to create a menu item across operating systems.

MenuItemNameDict

NameComplex

Platform

Platforms

UTTypeDeclarationModel

A model representing a Uniform Type Identifier (UTI) declaration.

Windows

Windows-specific instructions for menu item configuration.

Enums

CFBundleTypeRole

Environment

LSHandlerRank

NameField